Add boolean conceptual factorization
Use a simple set-covering approach to ensure every tight cross is
covered by concepts, pruning factorizations that can never be minimal
in the number of concepts. Exploit the fact that those concepts that
are simultaneously object concepts and attribute concepts must always be
contained in any factorization.

Note that pruning is still non-optimal since we never propagate the
updated bounds upwards, and thus still find non-minimal covers. In
practice, however, this turns out not to be a problem.

Signed-off-by: Maximilian Marx <mmarx@wh2.tu-dresden.de>